safezone/requirements.txt
asgiref==3.7.2
branca==0.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
Django==5.0
folium==0.15.1
idna==3.6
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.3
pillow==10.2.0
requests==2.31.0
sqlparse==0.4.4
tzdata==2024.1
urllib3==2.1.0
xyzservices==2023.10.1
